woot more tests

+++ mysimpledb/src/java/simpledb/BufferPool.java
@@ -114,6 +114,7 @@ public class BufferPool {
     public void releasePage(TransactionId tid, PageId pid) {
     	System.out.println("RELEASING PAGE!!!!!!!" + pid);
     	 Lock lock = lm.getLockTable().get(pid);
+    	 if (lock==null) {return;}
          lock.deleteTransaction(tid);
          if (lm.getLockTable().containsKey(pid)) {
         	 lm.getLockTable().remove(pid);
@@ -166,8 +167,10 @@ public class BufferPool {
     			System.out.println("committing... setting before image for page " + bpool.get(pgId));
     			// use current page contents as the before-image
     	        // for the next transaction that modifies this page.
-    	        Page p = bpool.get(pgId);
-    	        p.setBeforeImage();
+    			if (bpool.keySet().contains(pgId)) {
+    				Page p = bpool.get(pgId);
+    				p.setBeforeImage();
+    			}
     		}
     		System.out.println("in if commit");
     		flushPages(tid);
@@ -178,11 +181,13 @@ public class BufferPool {
     		System.out.println("size of list = " + list.size());
     		for (PageId pgID : list) {
     			
-    			System.out.println("bpool size = " + bpool.size());
-    			System.out.println("pageid? =" + pgID);
-    			bpool.get(pgID).markDirty(false, tid);
-    			bpool.remove(pgID);
-    			pageAccessTime.remove(pgID);
+    			//System.out.println("bpool size = " + bpool.size());
+    			//System.out.println("pageid? =" + pgID);
+    			if (bpool.keySet().contains(pgID)) {
+    				bpool.get(pgID).markDirty(false, tid);
+    				bpool.remove(pgID);
+    				pageAccessTime.remove(pgID);
+    			}
     			//releasePage(tid, pgID);
     		}
 
@@ -269,8 +274,13 @@ public class BufferPool {
      * cache.
      */
     public synchronized void discardPage(PageId pid) {
-    	if (bpool.contains(pid)){
+    	System.out.println("discarding page bpool size = " + bpool.size());
+    	System.out.println("first thing in bufferpool = " + bpool.keySet().toArray()[0]);
+    	if (bpool.keySet().contains(pid)){
+    		System.out.println("removing " + pid + "from bpool");
+    		System.out.println("size = " + bpool.size());
     		bpool.remove(pid);
+    		System.out.println("size = " + bpool.size());
     	}                                                                           // cosc460
     }
 
@@ -280,7 +290,7 @@ public class BufferPool {
      * @param pid an ID indicating the page to flush
      */
     private synchronized void flushPage(PageId pid) throws IOException {
-    	System.out.println("FLUSH PAGE");
+    	System.out.println("FLUSH PAGE where pid is" + pid);
     	// only call this on dirty page to evict
     	if (pid == null) {
     		throw new NullPointerException();
